Fixed pluralisation in notifications
